requirements.txt
flask
pymongo
wordcloud
pillow